ATA: SATA_MV: Fix probe failure when not phy driver
The generic phy layer does not cleannly support optional phys. It is
not possible to determine from the error code if there is expected to
be a phy according to DT, but it cannot be found, or no phy is listed
in DT. All that can be determined is that a phy is expected, but the
driver has not been loaded yet, in which case -EPROBE_DEFER is
returned. Thus play safe, consider all errors except -EPROBE_DEFER to
be none fatal and keep going, and in the case of -EPROBE_DEFER exit
the probe function with that error code.

Reported-by: Jean Pihet <jean.pihet@linaro.org>
Signed-off-by: Andrew Lunn <andrew@lunn.ch>