media: coda: jpeg: add CODA960 JPEG support
This patch adds JPEG encoding and decoding support for CODA960, handling
the JPEG hardware directly. Separate JPEG encoder and decoder video
devices are created due to the separate hardware unit and different
supported pixel formats. The hardware can not change subsampling on the
fly, but encode and decode 4:2:2 subsampled JPEG images from and into
this format.

There are two additional tracepoints added to the coda driver that can
be used together with the v4l2:v4l2_qbuf and v4l2:v4l2_dqbuf tracepoints
to to follow video frames through the mem2mem device when encoding or
decoding with the CODA960 JPEG codec:
    coda:coda_jpeg_run
    coda:coda_jpeg_done

Signed-off-by: Philipp Zabel <p.zabel@pengutronix.de>
[m.felsch@pengutronix.de: fix qsequence counting by explicitly
 checking for the !use_bit case]
Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>