net: dsa: mv88e6352: Enable forwarding of packets with unknown destination address
If a port is not member of a bridge group, offloaded or not, it should forward
all packets received from the CPU port to the designated output port, no matter
if the destination address of the packet is in its forwarding database or not.

If a port is member of a bridge group, offloaded or not, packets with unknown
destination address should be flooded to all members of the bridge group as part
of normal bridge operation.

In summary, there is no reason why flooding of packets with unknown destination
address should ever be disabled. This applies to unicast as well as multicast
packets.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>