bonding: Add a simple device tree binding
Some devices, like wireless access points, have two Ethernet
interfaces connected to an Ethernet switch embedded on the board.  To
make best use of these Ethernet ports, they need to be bonded in the
host, and a trunk created in the switch.

Since this is a property of the hardware, add a device tree binding to
the bond driver, so such hardware can be described via device tree.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>