bonding: Set protocol if bond interface used with DSA
eth_type_trans() will set skb->protocol to ETH_P_XDSA if the input
interface is being used for DSA. However when using bonding, the slave
interface is not used by DSA, so skb->protocol is left alone. When we
re-parent the skb from the slave to the bond interface, we should
re-evaluate the ether type and set it to ETH_P_XDSA if the bond
interface is being used by DSA. Calling eth_type_trans() here is not
possible because it has unwanted side effects.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>