ARM: dts: vf610-zii-dev: Add ZII development boards.
This commit adds support for both Rev. A and Rev. B of a Zodiac Inflight
Innovations development board, mainly intended for DSA development and
proof-of-concept testing.

The first generation of the board had several (hardware) issues, most of
which have been fixed with Rev. B.

The common parts are separated out into a .dtsi file, with individual
.dts files supporting Rev. A and Rev. B.

Signed-off-by: Cory Tusar <cory.tusar@pid1solutions.com>
Signed-off-by: Andrew Lunn <andrew@lunn.ch>