misc: eeprom: nameplate: Driver to export EEPROM contents in /sys
EEPROMs can be used to contain product information such as serial
number, date of manufacturing, modification history etc. If this
information can be expressed as nvmem elements, this driver allows
them to be exported in /sys. Elements can be exported as strings or
integers, as described in device tree.

NOT SIGNED OFF FOR MAINLINE