net: dsa: mv88e6xxx: isolate unbridged ports
As for the bridged ports, reserve a few VLANs to isolate ports that have
not been bridged yet.

By default, a port will be isolated with the CPU and DSA ports. When the
port joins a bridge, it will leave its reserved port. When it is removed
from a bridge, it will join its reserved VLAN again.

Fixes: 5fe7f68016ff ("net: dsa: mv88e6xxx: fix hardware bridging")
Signed-off-by: Vivien Didelot <vivien.didelot@savoirfairelinux.com>