net: dsa: Add parsing of an interrupt property
Some switch chips have an interrupt line, which can be connected to
a GPIO line and used to indicate interrupts. Add parsing of such
an interrupt to the DSA framework.

Also, update the binding, which already listed an interrupt.  However,
interrupts are a property of the chip, not the cluster, so move it
into the chip level of the hierarchy.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>