batman-adv: Handle parent interfaces in a different netns
batman-adv tries to prevent the user from placing a batX soft
interface into another batman mesh as a hard interface. It does this
by walking up the devices list of parents and ensures none batX
interfaces. However with the use of network name spaces, the parent
interface could be in a different name space. Remove the warning if
the parent interface cannot be found.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>