dsa: mv88e6xxx: Add support for 6390 family
The 6390 family has 10 external ports, and one internal port to a Z80
based processor.

It is similar to the 6352 family, so let that driver probe it. 6390
however has a different timer base for aging out entries from the
ATU. Add support for its 3.75s tick. It also places its port control
registers at a different MDIO address offset to all other Marvell
devices. Replace the hard coded REG_PORT() macro with a function to
return the correct MDIO address for a port.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>