net: dsa: Add basic switch register/unregister functions
A switch registers with the DSA core as part of its probe function.
The core needs to find the DSA Tree this switch belongs to, and create
it if it does not yet exist. The switch is then added to the tree.
The unregister will remove a switch from a tree, and if it is the last
switch, free the tree.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>