mv88e6xxx: Refactor MDIO to allow driver to register bus
Have the switch driver register its own MDIO bus. This allows for an
mdio-bus property in the device tree, with child nodes for phys, which
can be referenced via phandles, etc.

The common setup needs to be performed earlier, since the mutex will
be used when the MDIO bus is scanned on register.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>