net: dsa: mv88e6xxx: Support finding the mv88e6390 on the bus
The mv88e6390 places its port registers in a different place to all
other supported Marvell devices. The probe code looks for the device
ID in the first port. Hence the mv88e6390 is never detected. Add a new
compatibility string, and have the OF match data indicate the port
offset.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>