media: coda: allow to dump metas to the log and bitstream into a device coredump
If start_streaming is called with less than 512 bytes payload in the
bitstream ringbuffer, the decoder can not start. Dump the metas and the
first five bytes of each buffer to the log, and write the complete
bitstream into a device coredump for debugging purposes.
If the BIT processor PIC_RUN command times out, do the same.

Signed-off-by: Philipp Zabel <p.zabel@pengutronix.de>