Refactor reply reception logic
Take a first stab at simplifying MFD locking scheme. Convert the code to
only shar a pointer to a reply object, as opposed to have multiple
fileds protected by the lock. Also get rid of a spinlock since none of
the code should be ever called in atomic context.

TODO: Revisit the possiblity of getting rid of reply_lock