Refactor command generation and ditch hw_id
Provide 'struct of function pointers' as a device specific data instead
of singl numrical constant. Hopefully this would lead to a cleaner
codebase and allow us to corall most of hw differencies code into a
small set of functions.

This also allows us to ditch hw_id field and replace it with
zii_pic_variant struct.

As a part of this change command response generation was changed from
using hardcoded reply constants to calculated reply constants
implemneted by the meanst of 'translate' function