gpu: ipu-v3: ipu-dc: setup microcode from u64 array
Preparing the DC microcode templates on the stack simplifies the code
and allows to more clearly view the microcode that is actually written
to the microcode template memory.
The 42-bit DC microcode words are created from instruction macros with
their parameter names aligned to the documentation. Only the WRG and
WROD instrucion are used, so drop the WCLK instruction for now.

Signed-off-by: Philipp Zabel <p.zabel@pengutronix.de>