crypto: caam - add API to allocate Job Ring 0 especifically
By default, job ring 0 is the owner of the Secure Memory area
within CAAM, and therefore users (e.g. the sm_store driver) may want
to especifically allocate and use it.

This logic assumes that job ring devices are added to the list in
order, as it seems to be the case, so we just use the first one in the
list, regardless of how many other users (as given by the reference
count) it has.

Signed-off-by: Aleksander Morgado <aleksander@aleksander.es>