drm/etnaviv: properly implement mmaping of imported buffers
The intention of the existing code was to deflect the actual work
of mmaping a dma-buf to the exporter, as that one probably knows best
how to handle the buffer. Unfortunately the call to drm_gem_mmap did
more than what etnaviv needs in this case by actually setting up the
mapping.

Move mapping setup to the shm buffer type mmap implementation so we
only need to look up the BO and call the buffer type mmap function
from the handler.

Fixes mmap behavior with dma-buf imported and userptr buffers.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>