soc: imx: gpc: fix power up sequencing
Currently we wait only until the PGC inverts the isolation setting
before disabling the peripheral clocks. This doesn't ensure that the
reset is properly propagated through the peripheral devices in the
power domain.

Wait until the PGC singnals that the power up request is done and
wait a bit for resets to propagate before disabling the clocks.

Fixes GPU hangs at low devices temperature, due to incomplete
peripheral reset.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>