media: coda: normalise debug output
Consistently add the context index to debug output, which otherwise is
impossible to make sense of when two contexts are running concurrently.
For this purpose, add a convenience macro coda_dbg(). Use the function
name with the coda_ prefix stripped as keyword and vid-out and vid-cap
names for the queues. Add sequence counters to the decoder job finished
message and correctly indicate B frames. Add a start streaming message
to complement the stop streaming message.

Signed-off-by: Philipp Zabel <p.zabel@pengutronix.de>