zii: net: dsa: mv88e6xxx: add BPDU MAC to side ports
This patch puts the BPDU MAC address in the ATU for the netleft
and netright ports, so that the switch bypass BPDU frames.

This is RDU specific, so wrap this code in a new
NET_DSA_MV88E6XXX_ZII_RDU Kconfig symbol for tweaks touching the
driver setup itself.

Signed-off-by: Vivien Didelot <vivien.didelot@gmail.com>