drm/etnaviv: track previous gpu mmu context
When scheduling a new job we take a reference on the previous GPU MMU
context to avoid it being freed before the GPU has jumped to the
newly scheduled MMU context.

This keeps the previous context alive unit the new job has finished
execution, while technically it could be freed at the start of the
execution of the new job. Batching freeing of the previous context
with new job finish simplifies tracking and doesn't need any new
signals.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>