fs: super: Force sync prior to read-only remount to flush removed files.
Fix an -EBUSY seen when re-mounting a read-only JFS partition as
read-write, removing a file, and then re-mounting that same partition
again as read-only.  This was traced to a check in do_remount_sb() in
the upper filesystem layer for any open files which had been unlinked,
but not yet had that change flushed to disk.

NOTE - This is a Zodiac-specific change to eliminate a regression as
       compared to older 2.6.35.3 kernels used on MX51-based LRUs, and
       SHOULD NOT be upstreamed.

Signed-off-by: Cory Tusar <cory.tusar@zii.aero>