ARM: dts: imx6: RDU2: link eDP bridge to panel
This makes the eDP bridge chip aware of the panel, so it properly
advertises eDP instead of DP to userspace and allows for proper
backlight control through the panel abstraction. This patch is just
to add all the required nodes, actual linking of the display chain
will be done by the bootloader depending on panel detection.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>